Conflicts:

	compiler/cpp/src/generate/t_py_generator.cc

+++ lib/cpp/src/server/TNonblockingServer.h
@@ -1,8 +1,21 @@
-// Copyright (c) 2006- Facebook
-// Distributed under the Thrift Software License
-//
-// See accompanying file LICENSE or visit the Thrift site at:
-// http://developers.facebook.com/thrift/
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 
 #ifndef _THRIFT_SERVER_TNONBLOCKINGSERVER_H_
 #define _THRIFT_SERVER_TNONBLOCKINGSERVER_H_ 1
@@ -158,7 +171,7 @@ class TNonblockingServer : public TServer {
    *
    * @return the current limit on TConnection pool size.
    */
-  int getConnectionStackLimit() const {
+  size_t getConnectionStackLimit() const {
     return connectionStackLimit_;
   }
 
@@ -167,7 +180,7 @@ class TNonblockingServer : public TServer {
    *
    * @param sz the new limit for TConnection pool size.
    */
-  void setConnectionStackLimit(int sz) {
+  void setConnectionStackLimit(size_t sz) {
     connectionStackLimit_ = sz;
   }
